Improved RL Code
Improved RL Code (?)

Changed action space to be in range -255 to 255.

feature_grid how resizes image to 1080 by 1080 (square, probably should set it to be some parameter instead of hardcoding it).

"Reset" now adds noise to the feature grid when reset, allows for better exploration because initial point is random (you can remove it if you feel like it doesn't help).

Observation space now returns the image and the confidence in the target class.

Updated reward. Reward penalizes if top 1 is not target (-1), rewards the agent if it reaches christmas tree by the confidence of christmas tree x 10. Finally, penalizes agent for each step.

We are only concerned with a single noise, not accumulative noise. Replaced "env.feature_grid += noise". The image + current step generated noise is saved as adv image.

Implemented Boltzmann Exploration. Encourage exploration for a bit before switching to exploitation.